Misc Changes
Made misc changes within game, none of which were large enough to note

+++ Ludum Dare 46/Assets/Jade/BaseClasses/Entity.cs
@@ -5,7 +5,7 @@ public class Entity : MonoBehaviour, IDamageable
 {
     private int health = 3;
     protected int maxHealth = 10;
-    protected int speed = 2;
+    protected float speed = 2;
     protected Rigidbody rb;
 
     protected virtual void Start()
